Add files via upload
Instead of creating a conventional 'alphabet' counter array and then visiting every element to check whether it's zero or not. Here i have used only a number n where i subtract or add values in powers of 10 accordingly. For eg. a-1,b-10,c-100,d-1000,e-100000 etc.
in this case we only need to check whether n is equal to zero.